UDP connection in python
I added sample output code for the equipment number for client, but can't add anything currently until the database is up and running! Run server through VSCode and client in cmd. Ignore large commented out section in udpserver, that was me testing stuff but I left it in case it comes in handy later.